seperate liquidity management param controls
